requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.51.0